Fix votes on landing page
